Fix potential precision issue with division
Because the DatabaseSeeker is querying integer values which are then used for mathematical operations like division in some places, it is possible that different DBMS produce weird or wrong results for the division. As an example, the division of `1022/1023` is evaluated to `0` by SQL Server 2017 but to `0.9990` by MySQL and MariaDB. By casting the integers to floats prior to the division, this problem is solved.