Improve query performance through single-table index (#21)
* Add staudenmeir/laravel-cte

* Add migration for new scout_index table

The new indexing table has a combined structure similar to scout_documents and scout_words combined. The reason for this denormalized structure is that it will yield reduced concurrency issues for inserts.

* Implement new indexing strategy

The new simple strategy is delete-all-insert-all; while the final strategy could be delete-removed-insert-new. A test needs to be carried out later.

* Fix indexing

* Implement seeking for new schema

* Update tests

* Remove obsolete configuration

* Update README.md

* Fix migration for MySql

* Support tests for different database providers